% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two numbers together}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
A numeric value
}
\description{
Pretty cool function to add to numbers together
}
\examples{
add(7, 11)
add(15, 23)
}
