% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_wdi.R
\name{clean_wdi}
\alias{clean_wdi}
\title{clean_wdi()}
\usage{
clean_wdi(wdi_df)
}
\arguments{
\item{wdi_df}{data.frame: data returned by get_wdi()}
}
\value{
data.frame
}
\description{
Clean WDI data returned by get_wdi()
}

