% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_height.R
\name{cyl_height}
\alias{cyl_height}
\title{Calculate the height of a cylinder}
\usage{
cyl_height(radius, volume)
}
\arguments{
\item{radius}{Any number greater than zero}

\item{volume}{Any number greater than zero}
}
\value{
Volume divided by the product of \code{radius} and pi.
}
\description{
Calculate the height of a cylinder
}
\examples{
cyl_height(2,25)
cyl_height(5,1)

}
