% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_radius.R
\name{cyl_radius}
\alias{cyl_radius}
\title{Calculate the radius of a cylinder}
\usage{
cyl_radius(height, volume)
}
\arguments{
\item{height}{Any number greater than zero}

\item{volume}{Any number greater than zero}
}
\value{
Volume divided by the product of \code{height} and pi.
}
\description{
Calculate the radius of a cylinder
}
\examples{
cyl_radius(2,2.7)
cyl_radius(5,1)

}
