% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyl_volume.R
\name{cyl_volume}
\alias{cyl_volume}
\title{Calculate the volume of a cylinder}
\usage{
cyl_volume(radius, height)
}
\arguments{
\item{radius}{Any number greater than zero}

\item{height}{Any number greater than zero}
}
\value{
The product of \code{radius}, \code{height} and pi.
}
\description{
Calculate the volume of a cylinder
}
\examples{
cyl_volume(2,4)
cyl_volume(5,1)

}
