\name{fancy.sqrt}
\alias{fancy.sqrt}
\title{A function for ..........}
\description{
A description of the function ..............}
\usage{
fancy.sqrt(x)
}
\arguments{
  \item{x}{a numeric value.}
}
\value{
The function returns the square rooted value of its argument.
}

\examples{

# this should work
fancy.sqrt(10)
    
\dontrun{
# whereas this one shouldn't:
fancy.sqrt("character")
}
}
