% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_mean.R
\name{gamma_mean}
\alias{gamma_mean}
\title{Compute the mean of a Gamma(alpha, beta) random variable.}
\usage{
gamma_mean(alpha, beta)
}
\arguments{
\item{alpha}{The shape parameter of the Gamma distribution.}

\item{beta}{The rate parameter of the Gamma distribution.}
}
\value{
The mean of the Gamma variable.
}
\description{
Compute the mean of a Gamma(alpha, beta) random variable.
}
\examples{
gamma_mean(4,2)
}
