% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wdi.R
\name{get_wdi}
\alias{get_wdi}
\title{get_wdi()}
\usage{
get_wdi(wdi_code, start = 1960, end = 2015, remove_na = FALSE, ...)
}
\arguments{
\item{wdi_code}{character: The WDI code}

\item{start}{numeric: Start of the period}

\item{end}{numeric: End of the period}

\item{remove_na}{logical: Remove empty records if TRUE}

\item{...}{Other parameters to be passed to wbstats::wb()}
}
\value{
data.frame
}
\description{
Download WDI data using the WDI API
}

