% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sayhello.R
\name{hello}
\alias{hello}
\title{The hello function}
\usage{
hello(who)
}
\arguments{
\item{who}{a string for who will be greeted}
}
\value{
prints to screen
}
\description{
The hello function
}
\examples{
"Hello Isabel"

}
