% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kindoftrivial.R
\name{kindoftrivial}
\alias{kindoftrivial}
\title{A (Kind of) Trivial Function}
\usage{
kindoftrivial(x, y)
}
\arguments{
\item{x, y}{two values to be multiplied}
}
\value{
the product x*y
}
\description{
Just illustration; nothing really here
}
\examples{
trivial(2,3)
}

