% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightReadings.R
\docType{data}
\name{lightReadings}
\alias{lightReadings}
\title{Geolocator light intensity readings}
\format{
A data frame with 22320 rows and 2 variables:
\describe{
  \item{datetime}{A time series every two minutes from 2010-07-21 00:01:00}
  \item{light}{Light intensity measurements}
}
}
\usage{
data(lightReadings)
}
\description{
Light intensity measurements from a light logger commonly used for geolocation.
}
\keyword{datasets}
