% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/myFirstPackage.R
\docType{package}
\name{myFirstPackage}
\alias{myFirstPackage}
\alias{myFirstPackage-package}
\title{A toy package (TITLE)}
\description{
This does nothing. I mean nothing.
What're you looking at? (DESCRIPTION)
}
\details{
We can put more verbiage here if we want.
}

