% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgdraw_wrap.R
\name{PG_draw}
\alias{PG_draw}
\title{Wrapper for drawing from Pólya-Gamma distribution using the pgdraw package}
\usage{
PG_draw(b, c)
}
\arguments{
\item{b}{Shape parameter of the Pólya-Gamma - an integer.}

\item{c}{Second parameter of the Pólya-Gamma - a real.}
}
\value{
One draw from the Pólya-Gamma distribution.
}
\description{
Wrapper for drawing from Pólya-Gamma distribution using the pgdraw package
}
\examples{
x <- PG_draw(4,1)
x
}
