% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rectangle.R
\name{plot_rectangle}
\alias{plot_rectangle}
\title{Plot rectangle}
\usage{
plot_rectangle(height, width)
}
\arguments{
\item{height}{A number}

\item{width}{A number}
}
\value{
Plots a rectangle of the given height and width.
}
\description{
Plot rectangle
}
\examples{
plot_rectangle(2,3)
plot_rectangle(5,3)

}
