% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wdi.R
\name{plot_wdi}
\alias{plot_wdi}
\title{plot_wdi()}
\usage{
plot_wdi(wdi_df)
}
\arguments{
\item{wdi_df}{data.frame: data returned by clean_wdi()}
}
\value{
list
}
\description{
plot WDI data returned by clean_wdi()
}

