% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCircle.R
\name{plotCircle}
\alias{plotCircle}
\title{Plot a circle of radius R.}
\usage{
plotCircle(...)
}
\arguments{
\item{r}{The radius of the circle to be plotted}
}
\value{
None  The function is called solely for its side effect
}
\description{
Input the radius r and get a
plot of a circle with a solid line
}
\examples{
plotCircle(10)
}

