% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfmodel.R
\docType{data}
\name{rfmodel}
\alias{rfmodel}
\title{Pretrained ML model}
\format{
\subsection{\code{rfmodel}}{

A random forest model
To use the random forest, the user needs to prepare 6 columns
prepare mz0, int2_o_int0, int1_o_int0, RI2_RI1, mz_2_0, mz_1_0
}
}
\source{
pubchem
}
\usage{
rfmodel
}
\description{
A model trained from PubChem
}
\keyword{datasets}
