% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split a string into parts}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector (length 1)}

\item{pattern}{What to split on}

\item{n}{Max number of pieces}
}
\value{
A character vector
}
\description{
Split a string into parts
}
\examples{
str_split_one("a,b,c", ",")
}
