% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{sum.matrix}
\alias{sum.matrix}
\title{Sum the elements of a matrix}
\usage{
\method{sum}{matrix}(X, n_threads = 4)
}
\arguments{
\item{X}{A matrix}

\item{n_threads}{The number of threads to use}
}
\value{
The sum of the elements of the matrix
}
\description{
Sum the elements of a matrix
}
\examples{
X <- matrix(1:10, 2, 5)
sum_matrix(X)
}
