% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sum_matrix}
\alias{sum_matrix}
\title{Sum the elements of a matrix}
\usage{
sum_matrix(X, n_threads = 4L)
}
\arguments{
\item{X}{A matrix}
}
\value{
The sum of the elements of the matrix
}
\description{
Sum the elements of a matrix
}
\examples{
X <- matrix(1:10, 2, 5)
sum_matrix(X)
}
