% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\title{Compute Variance}
\usage{
variance(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Compute Variance for a numeric vector
}
\examples{
x <- rnorm(100,3,3)
print(variance(x))
}


