% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greetings.R
\name{greetings}
\alias{greetings}
\title{greeting}
\usage{
greetings(many_names)
}
\arguments{
\item{many_names}{Several names}
}
\value{
A greeting
}
\description{
greeting
}
\examples{
greetings(c("Alice", "Bob"))
}
