% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCircle.R
\name{plotCircle}
\alias{plotCircle}
\title{Plot a circle of radius r.}
\usage{
plotCircle(r = 1)
}
\arguments{
\item{r}{the radius of the circle to be plotted.}
}
\value{
None the function is called solely for its side effect
}
\description{
Normally, we would now describe what
 the function is supposed to do.
}
\examples{
plotCircle(10)
}

