% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Splits a string}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector with one element.}

\item{pattern}{What to split on.}

\item{n}{Maximum number of pieces to return.}
}
\value{
A character vector
}
\description{
Splits a string
}
\examples{
x <- "a,b,c"
str_split_one(x, pattern = ",")
}
